Handle ACU termination gracefully
If you were to stop the ACU process outside of supervision (i.e. `GenServer.stop(acu)`)
it will leave the Transport connection and Circuits.UART reference open and orphaned.
This causes issues later on when attempting to start the ACU again.

This change adds handling so that gracefully terminating the ACU process will propgate
to the Transport and Circuits.UART processes as well