Fix requests to offline devices
This is for situations where the device is registered,
but is offline (possibly not powered up or addressed correctly yet)

In those cases, a command sent for the device would never get processed
and this would subsequently fail with a GenServer timeout.
This is slightly misleading and could cause other process failures.

The changes here fix that by removing the epectation that the first
command to a device would always be a `POLL`. Instead, whatever command
is next in the queue would be sent with sequence 0 and would either
establish communication or fail with a timeout that is handled in the ACU
process.

After reading the spec a little closer and testing, there does not seem
to be a requirement to sequence 0 always being POLL, so this should be
a safe change.