Allow specifying timeout for commands
Previously the timeout was hardcoded to the max allowed reply delay
(200ms) for every command. There may be some cases where you want to
expect data sooner and need to be able to specify a longer timeout.

This change adds a `:timeout` option avaialable to all commands with
protections that the value is > 0 and <= 200 else the default is used